+++ src/plugins/nonebot_plugin_larkutils/url_validator.py
