+++ src/plugins/nonebot_plugin_preview/__init__.py
@@ -11,6 +11,5 @@ require("nonebot_plugin_larklang")
 require("nonebot_plugin_htmlrender")
 require("nonebot_plugin_larkutils")
 require("nonebot_plugin_alconna")
-require("nonebot_plugin_chat")
 
 from . import __main__
